requirements.txt
altgraph==0.17.2
certifi==2021.10.8
charset-normalizer==2.0.10
colorama==0.4.4
decorator==4.4.2
future==0.18.2
idna==3.3
imageio==2.13.5
imageio-ffmpeg==0.4.5
moviepy==1.0.3
numpy==1.22.0
opencv-python==4.5.5.62
openrgb-python==0.2.15
pefile==2021.9.3
Pillow==9.0.0
proglog==0.1.9
pyinstaller==4.8
pyinstaller-hooks-contrib==2021.5
pywal==3.3.0
pywin32-ctypes==0.2.0
requests==2.27.1
tqdm==4.62.3
urllib3==1.26.8
